+++ lib/src/core/extensions/boolean_extension.dart
@@ -1,0 +1,5 @@
+extension NullBool on bool? {
+  bool get isTrue => this == true;
+
+  bool get isFalse => this == false;
+}